requirements.txt
fastapi
uvicorn
transformers
torch
psycopg2-binary
kafka-python
jinja2
pdfkit
